refactoring - custom icons are now only managed by the icon-dictionary service
